Minor changes in repos and services

+++ src/com/elena/TravelAgency/v3/City/service/CityMemoryService.java
@@ -4,25 +4,30 @@ import com.elena.TravelAgency.v3.City.City;
 import com.elena.TravelAgency.v3.City.repo.CityMemoryRepo;
 
 public class CityMemoryService implements CityService {
-    private CityMemoryRepo cityMemoryRepo = new CityMemoryRepo();
+    private final CityMemoryRepo cityMemoryRepo = new CityMemoryRepo();
 
     public void add(City city) {
-        cityMemoryRepo.add(city);
+        if (city != null)
+            cityMemoryRepo.add(city);
     }
 
     public void delete(Long id) {
-        cityMemoryRepo.delete(id);
+        if ( id != null)
+            cityMemoryRepo.delete(id);
     }
 
     public void delete(City city) {
-        cityMemoryRepo.delete(city);
+        if (city != null)
+            cityMemoryRepo.delete(city);
     }
 
     public void find(Long id) {
-        cityMemoryRepo.find(id);
+        if (id != null)
+            cityMemoryRepo.find(id);
     }
 
     public void find(String name) {
-        cityMemoryRepo.find(name);
+        if (name != null)
+            cityMemoryRepo.find(name);
     }
 }
